+++ src/main/java/com/toy4/domain/schedule/dto/response/ScheduleDayOffResponse.java
@@ -1,4 +1,4 @@
-package com.toy4.domain.dayOffHistory.dto.response;
+package com.toy4.domain.schedule.dto.response;
 
 import com.toy4.domain.dayOffHistory.domain.DayOffHistory;
 import lombok.AllArgsConstructor;
